Fixed dealing issue
dealCard method grabs a card from the back of the deck because the topcard is in the back when the deck is faced down.

The loop was taking the card from the top and was checking each card from the bottom. So, at the halfway point the checker would return with errorflags.
Fixed the loop so that it now checks the cards from the top